Switch from DecimalField to IntegerField

+++ platalbank_core/models/account.py
@@ -3,7 +3,7 @@ from rest_framework import serializers
 from rest_framework import viewsets
 
 class Account(models.Model):
-    balance = models.DecimalField(max_digits=20, decimal_places=2)
+    balance = models.IntegerField() # Number of cents owned
     description = models.CharField(max_length=1024)
     short_name = models.CharField(max_length=128)
     # TODO: owner = models.ForeignKey("LegalPerson")
